final version of collection entry type definition - the test phase of the new collection data structure is finished - test data that had been generated is void. There will be no migration - the new collection files are located in DATA/INDEX/PUBLIC/TEXT/RICOLLECTION - the index dump is void. There will be no migration - the new index dump is in DATA/INDEX/PUBLIC/TEXT/RICACHE
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2983 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/index/indexCachedRI.java
@@ -86,10 +86,6 @@ public class indexCachedRI implements indexRI {
         return entries.updated();
     }
     
-    public indexContainer emptyContainer(String wordHash) {
-        return new indexContainer(wordHash, payloadrow);
-    }
-    
     public indexContainer addEntry(String wordHash, indexRWIEntry entry, long updateTime, boolean intern) {        
         // add the entry
         if (intern) {
@@ -219,10 +215,9 @@ public class indexCachedRI implements indexRI {
     }
 
     public indexContainer deleteContainer(String wordHash) {
-        indexContainer c = new indexContainer(wordHash, payloadrow);
-        c.add(riIntern.deleteContainer(wordHash), -1);
-        c.add(riExtern.deleteContainer(wordHash), -1);
-        c.add(backend.deleteContainer(wordHash), -1);
+        indexContainer c = riIntern.deleteContainer(wordHash);
+        if (c == null) c = riExtern.deleteContainer(wordHash); else c.add(riExtern.deleteContainer(wordHash), -1);
+        if (c == null) c = backend.deleteContainer(wordHash); else c.add(backend.deleteContainer(wordHash), -1);
         return c;
     }
